Fixed getPhotosBeforeOrAfterTime() so that the unixTimeInSecs field is a double.

+++ fluxtream-web/src/main/java/com/fluxtream/api/BodyTrackController.java
@@ -672,7 +672,7 @@ public class BodyTrackController {
     public String getPhotosBeforeOrAfterTime(@PathParam("UID") long uid,
                                              @PathParam("ConnectorPrettyName") String connectorPrettyName,
                                              @PathParam("ObjectTypeName") String objectTypeName,
-                                             @PathParam("unixTime") long unixTimeInSecs,
+                                             @PathParam("unixTime") double unixTimeInSecs,
                                              @PathParam("count") int desiredCount,
                                              @QueryParam("isBefore") boolean isGetPhotosBeforeTime,
                                              @QueryParam("tags") List<String> tags,
@@ -688,7 +688,7 @@ public class BodyTrackController {
                 return gson.toJson(new StatusModel(false, "Invalid User ID (null)"));
              }
 
-            final SortedSet<PhotoService.Photo> photos = photoService.getPhotos(uid, unixTimeInSecs * 1000, connectorPrettyName, objectTypeName, desiredCount, isGetPhotosBeforeTime);
+            final SortedSet<PhotoService.Photo> photos = photoService.getPhotos(uid, (long)(unixTimeInSecs * 1000), connectorPrettyName, objectTypeName, desiredCount, isGetPhotosBeforeTime);
 
             // create the JSON response
             final List<PhotoItem> photoItems = new ArrayList<PhotoItem>();
@@ -715,9 +715,9 @@ public class BodyTrackController {
         long id;
         String description;
         String comment;
-        long begin_d;
+        double begin_d;
         String begin;
-        long end_d;
+        double end_d;
         String end;
         String dev_id;
         String dev_nickname;
@@ -735,7 +735,7 @@ public class BodyTrackController {
             this.id = photoFacetVO.id;
             this.description = photoFacetVO.description == null ? "" : photoFacetVO.description;
             this.comment = photoFacetVO.comment == null ? "" : photoFacetVO.comment;
-            this.begin_d = photoFacetVO.start / 1000; // convert millis to seconds
+            this.begin_d = photoFacetVO.start / 1000.0; // convert millis to seconds
             this.begin = DATE_TIME_FORMATTER.print(photoFacetVO.start);
             this.end_d = this.begin_d;
             this.end = this.begin;
